# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Barbershop API
  version: 1.0.0
  description: |
    Backend API for barbershop booking.
    - JWT Bearer auth for /me, /bookings, /admin/*
    - Public endpoints for barbers/services/availability
servers:
  - url: http://localhost:{port}/api
    description: Local dev
    variables:
      port:
        default: "3000"
...
paths:
...
  /admin/services:
    get:
      tags: [Admin - Services]
      summary: List services (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Services
          content:
            application/json:
              schema:
                type: object
                properties:
                  services:
                    type: array
                    items:
                      $ref: "#/components/schemas/Service"
                required: [services]
    post:
      tags: [Admin - Services]
      summary: Create service
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  service:
                    $ref: "#/components/schemas/Service"
                required: [service]
        "409":
          description: Service name already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/services/{id}:
    patch:
      tags: [Admin - Services]
      summary: Update service
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/ObjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceUpdateRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  service:
                    $ref: "#/components/schemas/Service"
                required: [service]
        "409":
          description: Service name already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags: [Admin - Services]
      summary: Delete service
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/ObjectId"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    $ref: "#/components/schemas/Service"
                required: [deleted]
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ObjectId:
      type: string
      description: Mongo ObjectId
      pattern: "^[0-9a-fA-F]{24}$"
      example: "650f1c2b3a4d5e6f7b8c9d01"
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Service:
      type: object
      properties:
        _id: { $ref: "#/components/schemas/ObjectId" }
        name: { type: string }
        durationMin: { type: integer, minimum: 1 }
        price: { type: number, minimum: 0 }
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [_id, name, durationMin, price]
    ServiceCreateRequest:
      type: object
      properties:
        name: { type: string }
        durationMin: { type: integer, minimum: 1 }
        price: { type: number, minimum: 0 }
      required: [name, durationMin, price]
    ServiceUpdateRequest:
      type: object
      properties:
        name: { type: string }
        durationMin: { type: integer, minimum: 1 }
        price: { type: number, minimum: 0 }
...